+++ DesignerMacros/src/services-cqrs-crud/query-on-map/query-on-map.ts
@@ -1,7 +1,7 @@
 /// <reference path="../common/query-on-map.ts" />
 
 /**
- * Used by Intent.Modules\Modules\Intent.Modules.Metadata.DocumentDB
+ * Used by Intent.Modules\Modules\Intent.Modules.Application.MediatR.CRUD
  *
  * Source code here:
  * https://github.com/IntentArchitect/Intent.Modules/blob/master/DesignerMacros/src/serivces-cqrs-crud/query-on-map/query-on-map.ts