+++ DesignerMacros/src/services-rest-route-validation/operation/operation-rest-route-validation.ts
@@ -13,6 +13,7 @@ function validateRestRoutesOperation(element: MacroApi.Context.IElementApi): Str
     return validateRestRoutes(element);
 }
 
+//Comment / UnComment below when you publish
 validateRestRoutesOperation(element);
 //return validateRestRoutesOperation(lookup(id));